prevent state loop
